requirements.txt
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.5.0
moviepy>=1.0.3,<2.0.0
Pillow>=10.0.0
edge-tts>=6.1.0
requests>=2.31.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
python-dotenv>=1.0.0
tenacity>=8.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
